+++ src/decorators/service.decorator.ts
@@ -8,9 +8,9 @@ import { Container } from '../container.class';
  * Marks class as a service that can be injected using Container.
  */
 /* eslint-disable-next-line @typescript-eslint/no-unused-vars */
-export function Service<T = unknown>(): Function;
-export function Service<T = unknown>(options: ServiceOptions<T>): Function;
-export function Service<T>(options: ServiceOptions<T> = {}): ClassDecorator {
+export function service<T = unknown>(): Function;
+export function service<T = unknown>(options: ServiceOptions<T>): Function;
+export function service<T>(options: ServiceOptions<T> = {}): ClassDecorator {
   return targetConstructor => {
     const serviceMetadata: ServiceMetadata<T> = {
       id: options.id || targetConstructor,